pyproject.toml
[tool.poetry]
name = "spbgti-tools-schedule"
version = "1.0.0"
description = ""
authors = ["Boger <kotvberloge@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
django = "==2.2.*,>=2.2.4"
gunicorn = "^19.9"
psycopg2-binary = "^2.8"
dj-database-url = "^0.5.0"
django_nose = "^1.4.6"
djangorestframework = "^3.10.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
